requirement.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
transformers
